productivity tool: finish the t-ssr implementation

+++ t-ssr/productivity-tool/src/Controller/MainController.php
@@ -33,7 +33,10 @@ class MainController extends AbstractController
     #[Route('/authenticate', name: 'authenticate')]
     public function auth(Request $request): Response
     {
-        $this->denyAccessIfAuthenticated();
+        $redirect = $this->denyAccessIfAuthenticated();
+        if ($redirect instanceof Response) {
+            return $redirect;  // Perform the redirect if necessary
+        }
 
         if ($request->isMethod('POST')) {
             $redirectRouteName = $this->dataService->authenticate($request->request);
@@ -48,7 +51,10 @@ class MainController extends AbstractController
     #[Route('/logout', name: 'logout')]
     public function logout(): Response
     {
-        $this->denyAccessUnlessAuthenticated();
+        $redirect = $this->denyAccessUnlessAuthenticated();
+        if ($redirect instanceof Response) {
+            return $redirect;  // Perform the redirect if necessary
+        }
         $redirectRouteName = $this->dataService->logout();
         if ($redirectRouteName) {
             return $this->redirectToRoute($redirectRouteName);
@@ -59,7 +65,10 @@ class MainController extends AbstractController
     #[Route('/todos', name: 'todos')]
     public function todos(Request $request): Response
     {
-        $this->denyAccessUnlessAuthenticated();
+        $redirect = $this->denyAccessUnlessAuthenticated();
+        if ($redirect instanceof Response) {
+            return $redirect;  // Perform the redirect if necessary
+        }
         return $this->render('main/todos.html.twig', $this->dataService->getTodosData(
             $request->query->get('lastPage', 1),
             $request->query->get('includeCompleted', false)
@@ -69,7 +78,10 @@ class MainController extends AbstractController
     #[Route('/todos/add', name: 'todos_add', methods: ['POST'])]
     public function addTodo(Request $request): Response
     {
-        $this->denyAccessUnlessAuthenticated();
+        $redirect = $this->denyAccessUnlessAuthenticated();
+        if ($redirect instanceof Response) {
+            return $redirect;  // Perform the redirect if necessary
+        }
         $flash = $this->dataService->addTodo($request->request);
 
         $this->addFlash($flash['type'], $flash['message']);
@@ -82,7 +94,10 @@ class MainController extends AbstractController
     #[Route('/todos/update', name: 'todos_update', methods: ['POST'])]
     public function updateTodo(Request $request): Response
     {
-        $this->denyAccessUnlessAuthenticated();
+        $redirect = $this->denyAccessUnlessAuthenticated();
+        if ($redirect instanceof Response) {
+            return $redirect;  // Perform the redirect if necessary
+        }
         $flash = $this->dataService->updateTodo($request->request->all());
 
         $this->addFlash($flash['type'], $flash['message']);
@@ -95,7 +110,10 @@ class MainController extends AbstractController
     #[Route('/todos/delete', name: 'todos_delete', methods: ['POST'])]
     public function deleteTodo(Request $request): Response
     {
-        $this->denyAccessUnlessAuthenticated();
+        $redirect = $this->denyAccessUnlessAuthenticated();
+        if ($redirect instanceof Response) {
+            return $redirect;  // Perform the redirect if necessary
+        }
         $flash = $this->dataService->deleteTodo($request->request->get('id'));
 
         $this->addFlash($flash['type'], $flash['message']);
@@ -108,7 +126,10 @@ class MainController extends AbstractController
     #[Route('/todos/add-spent-time', name: 'todos_add_spent_time', methods: ['POST'])]
     public function addSpentTime(Request $request): Response
     {
-        $this->denyAccessUnlessAuthenticated();
+        $redirect = $this->denyAccessUnlessAuthenticated();
+        if ($redirect instanceof Response) {
+            return $redirect;  // Perform the redirect if necessary
+        }
         $flash = $this->dataService->addSpentTime($request->request->get('todoId'), $request->request->get('duration'));
 
         $this->addFlash($flash['type'], $flash['message']);
@@ -121,7 +142,10 @@ class MainController extends AbstractController
     #[Route('/todos/delete-spent-time', name: 'todos_delete_spent_time', methods: ['POST'])]
     public function deleteSpentTime(Request $request): Response
     {
-        $this->denyAccessUnlessAuthenticated();
+        $redirect = $this->denyAccessUnlessAuthenticated();
+        if ($redirect instanceof Response) {
+            return $redirect;  // Perform the redirect if necessary
+        }
         $flash = $this->dataService->deleteSpentTime($request->request->get('id'));
 
         $this->addFlash($flash['type'], $flash['message']);
@@ -134,14 +158,20 @@ class MainController extends AbstractController
     #[Route('/todo-categories', name: 'todo_categories')]
     public function todoCategories(): Response
     {
-        $this->denyAccessUnlessAuthenticated();
+        $redirect = $this->denyAccessUnlessAuthenticated();
+        if ($redirect instanceof Response) {
+            return $redirect;  // Perform the redirect if necessary
+        }
         return $this->render('main/todo-categories.html.twig', $this->dataService->getTodoCategoriesData());
     }
 
     #[Route('/todo-categories/add', name: 'todo_categories_add', methods: ['POST'])]
     public function addTodoCategory(Request $request): Response
     {
-        $this->denyAccessUnlessAuthenticated();
+        $redirect = $this->denyAccessUnlessAuthenticated();
+        if ($redirect instanceof Response) {
+            return $redirect;  // Perform the redirect if necessary
+        }
         $flash = $this->dataService->addTodoCategory($request->request);
 
         $this->addFlash($flash['type'], $flash['message']);
@@ -152,7 +182,10 @@ class MainController extends AbstractController
     #[Route('/todo-categories/update', name: 'todo_categories_update', methods: ['POST'])]
     public function updateTodoCategory(Request $request): Response
     {
-        $this->denyAccessUnlessAuthenticated();
+        $redirect = $this->denyAccessUnlessAuthenticated();
+        if ($redirect instanceof Response) {
+            return $redirect;  // Perform the redirect if necessary
+        }
         $flash = $this->dataService->updateTodoCategory($request->request);
 
         $this->addFlash($flash['type'], $flash['message']);
@@ -163,7 +196,10 @@ class MainController extends AbstractController
     #[Route('/todo-categories/delete', name: 'todo_categories_delete', methods: ['POST'])]
     public function deleteTodoCategory(Request $request): Response
     {
-        $this->denyAccessUnlessAuthenticated();
+        $redirect = $this->denyAccessUnlessAuthenticated();
+        if ($redirect instanceof Response) {
+            return $redirect;  // Perform the redirect if necessary
+        }
         $flash = $this->dataService->deleteTodoCategory($request->request->get('id'));
 
         $this->addFlash($flash['type'], $flash['message']);
@@ -174,21 +210,28 @@ class MainController extends AbstractController
     #[Route('/spent-times', name: 'spent_times')]
     public function spentTimes(): Response
     {
-        $this->denyAccessUnlessAuthenticated();
+        $redirect = $this->denyAccessUnlessAuthenticated();
+        if ($redirect instanceof Response) {
+            return $redirect;  // Perform the redirect if necessary
+        }
         return $this->render('main/spent-times.html.twig', $this->dataService->getSpentTimesData());
     }
 
-    private function denyAccessUnlessAuthenticated(): void
+    private function denyAccessUnlessAuthenticated(): ?Response
     {
         if (!$this->requestStack->getSession()->get('is_authenticated')) {
-            $this->redirectToRoute('authenticate');
+            return $this->redirectToRoute('authenticate');
         }
+
+        return null;
     }
 
-    private function denyAccessIfAuthenticated(): void
+    private function denyAccessIfAuthenticated(): ?Response
     {
         if ($this->requestStack->getSession()->get('is_authenticated')) {
-            $this->redirectToRoute('home');
+            return $this->redirectToRoute('home');
         }
+
+        return null;
     }
 }
